third_party/WebKit/LayoutTests/fast/html/a-click.html

<!DOCTYPE HTML>
<body/>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<a id="target" href="#abc" title="This should appear as a tooltip">Hello.</a>
<script>
function doubleClickNoNavigation() {
  document.getElementById('target').dispatchEvent(new MouseEvent("click", {detail: 2}));
  assert_equals(window.location.hash, "", "There should be no navigation on double click");
}

test(doubleClickNoNavigation, "Double Click No Navigation");
</script>
